rename: 폴더명 Detail -> ProductDetail 로 변경
