+++ kakao-shop/src/components/page/ProductDetail/SelectOptionItemList/index.tsx
@@ -1,7 +1,7 @@
 import { Fragment, MouseEventHandler } from 'react';
 import type { UserSelectOption } from 'types/product';
 
-import SelectOptionItem from '@components/page/Detail/SelectOptionItem';
+import SelectOptionItem from '@components/page/ProductDetail/SelectOptionItem';
 
 type Props = {
   options?: UserSelectOption[];
